moving config.js content to new file... fixing imports as needed
